% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errantabun.R
\name{errantabun}
\alias{errantabun}
\title{Errant Abundance Matrix Function}
\usage{
errantabun(abun_matrix, method, prob)
}
\arguments{
\item{abun_matrix}{The original abundance matrix. Columns must represent species, and rows must represent plots/blocks/etc.}

\item{method}{The statistical distribution of choice: "Poisson" or "binomial".}

\item{prob}{The probability weight used to construct the binomial distribution. Not relevant for Poisson distribution.}
}
\description{
This function allows one to create an errant abundance matrix using an original abundance matrix and a random statistical distribution of choice.
}
\keyword{Poisson,}
\keyword{abundance,}
\keyword{binomial,}
\keyword{distribution,}
\keyword{errant,}
\keyword{errantabun()}
\keyword{random}
