% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksr.R
\docType{data}
\name{ksr}
\alias{ksr}
\title{Species abundance matrix from Koffler Scientific Reserve}
\format{
A matrix with 56 rows and 71 variables.
}
\source{
Leila Forsyth. Department of Ecology and Evolutionary Biology (University of Toronto). TBD

LazyData: true
}
\usage{
ksr
}
\description{
This species abundance matrix is collected from the Koffler Scientific Reserve at Joker's Hill (University of Toronto).

Fourteen blocks serve as disparate communities of grassland plant species. Each block contains an array of different-sized plots: eight 0.5m^2, eight 1m^2, and four 2m^2 plots.

Each column of this abundance matrix represents a grassland plant species. Each row represents one of four 2m^2 plots from all 14 blocks.
}
\references{
Leila Forsyth. Department of Ecology and Evolutionary Biology (University of Toronto). TBD
}
\keyword{abundance,}
\keyword{dataset,}
\keyword{matrix}
\keyword{species,}
