% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errantabun.R
\name{errantabun}
\alias{errantabun}
\title{Errant abundance matrix function}
\usage{
errantabun(abun, method, prob)
}
\arguments{
\item{abun}{The original abundance matrix. Columns must represent species, and rows must represent plots/blocks/etc.}

\item{method}{The statistical distribution of choice: "Poisson" or "binomial".}

\item{prob}{The probability weight used to construct the binomial distribution. Not relevant for Poisson distribution.}
}
\description{
`errantabun()` allows for the calculation of an errant species abundance matrix
using an original species abundance matrix and a statistical distribution of choice.

`errantabun()` utilizes the rpois and the rbinom functions to create an abundance
distribution for each species. The lambda for the Poisson distribution is the original
abundance (vector) for each species. The size for the binomial distribution is the original
abundance (vector) for each species.

`errantabun()` takes in a matrix as input and provides a data frame as output.
}
\keyword{Poisson,}
\keyword{abundance,}
\keyword{binomial,}
\keyword{distribution,}
\keyword{errant,}
\keyword{random}
