% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksr_species.R
\docType{data}
\name{ksr_species}
\alias{ksr_species}
\title{Species abundance dataset from KSR}
\format{
A dataset with 280 rows and 74 variables:
\describe{
 \item{Block}{block number, ranges from 1-14}
 \item{Plot}{plot number, ranges from 1-8}
 \item{Plot_size}{size of plot, in m^2}
 \item{species}{scientific name of species}
}
}
\source{
Leila Forsyth. Department of Ecology and Evolutionary Biology (University of Toronto). TBD

LazyData: true
}
\usage{
ksr_species
}
\description{
This species abundance dataset is collected from the Koffler Scientific Reserve at Joker's Hill (University of Toronto).

Fourteen blocks serve as disparate communities of grassland plant species. Each block contains an array of different-sized plots: eight 0.5m^2, eight 1m^2, and four 2m^2 plots.

This abundance dataset must be transformed into an abundance matrix, with columns representing species and rows representing the plot/block/etc. of choice.
}
\examples{
ksr_species_matrix <- ksr_species \%>\%
 filter(Size == 2) \%>\%
 unite(Plot, c(Block, Plot), remove = TRUE, sep = ".") \%>\%
 transform(Plot = as.numeric(Plot)) \%>\%
 column_to_rownames(var = "Plot") \%>\%
 data.matrix

}
\references{
Leila Forsyth. Department of Ecology and Evolutionary Biology (University of Toronto). TBD
}
\keyword{abundance}
\keyword{dataset,}
\keyword{species,}
